:recycle: User 로그인 Dto 및 Service, Controller 수정

+++ src/main/java/backend/whereIsMyTeam/user/dto/UserLoginResponseDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserLoginResponseDto {
-    private Long userIdx;
+    private Integer userIdx;
     private String token;
     private String refreshToken;
 }
